remove code-input
